src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-contents-select-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: target.createShadowRoot is not a function
